api/requirements.txt
# ── SamiX FastAPI Backend ──
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# AI & Language Models
groq>=0.9.0

# Speech-to-Text
deepgram-sdk>=3.2.0
openai-whisper>=20231117

# Audio
pydub>=0.25.1
audioop_lts

# RAG Stack
langchain>=0.2.5
langchain-community>=0.2.5
langchain-text-splitters>=0.2.2
langchain-huggingface>=0.0.3
sentence-transformers>=3.0.0
pymilvus>=2.5.3
langchain-milvus>=0.1.7
pymilvus[milvus_lite]
rank_bm25>=0.2.2

# Document Parsing
pypdf>=4.2.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Utilities
python-dotenv>=1.0.1
httpx>=0.27.0
python-multipart>=0.0.7
